Refactor test structure for improved modularity and dynamic testing: - Moved tests to individual files for better organization - Implemented temporary files and directories for dynamic testing - Integrated tests/data files for ease of use - Included an __init__.py file, that is empty, to allow for the import of the data directory as a module
